+++ app/components/primer/hidden_text_expander.rb
@@ -1,0 +1,43 @@
+# frozen_string_literal: true
+
+module Primer
+  # Use HiddenTextExpander to indicate and toggle hidden text.
+  class HiddenTextExpander < Primer::Component
+    # @example Default
+    #   <%= render(Primer::HiddenTextExpander.new) %>
+    #
+    # @example Inline
+    #   <%= render(Primer::HiddenTextExpander.new(inline: true)) %>
+    #
+    # @example Styling the button
+    #   <%= render(Primer::HiddenTextExpander.new(button_arguments: { p: 1, classes: "my-custom-class" })) %>
+    #
+    # @param inline [Boolean] Whether or not the expander is inline.
+    # @param button_arguments [Hash] <%= link_to_system_arguments_docs %> for the button element.
+    # @param system_arguments [Hash] <%= link_to_system_arguments_docs %>
+    def initialize(inline: false, button_arguments: {}, **system_arguments)
+      @system_arguments = system_arguments
+      @system_arguments[:tag] ||= :span
+      @system_arguments[:classes] = class_names(
+        "hidden-text-expander",
+        @system_arguments[:classes],
+        "inline" => inline
+      )
+
+      @button_arguments = button_arguments
+      @button_arguments[:tag] = :button
+      @button_arguments[:type] = :button
+      @button_arguments[:"aria-expanded"] = false
+      @button_arguments[:classes] = class_names(
+        "ellipsis-expander",
+        button_arguments[:classes]
+      )
+    end
+
+    def call
+      render(Primer::BaseComponent.new(**@system_arguments)) do
+        render(Primer::BaseComponent.new(**@button_arguments)) { "&hellip;" }
+      end
+    end
+  end
+end